Issue #4257, #4279: Move intent related code from browser-session to feature-intent.
