+++ components/feature/intent/src/main/java/mozilla/components/feature/intent/ext/IntentExtensions.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.browser.session.intent
+package mozilla.components.feature.intent.ext
 
 import android.content.Intent
 import mozilla.components.support.utils.SafeIntent
