+++ components/feature/intent/src/test/java/mozilla/components/feature/intent/processing/TabIntentProcessorTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.feature.intent
+package mozilla.components.feature.intent.processing
 
 import android.content.Intent
 import androidx.test.ext.junit.runners.AndroidJUnit4
