+++ components/feature/pwa/build.gradle
@@ -34,6 +34,7 @@ dependencies {
     implementation project(':concept-engine')
     implementation project(':concept-fetch')
     implementation project(':feature-customtabs')
+    implementation project(':feature-intent')
     implementation project(':feature-session')
     implementation project(':support-base')
     implementation project(':support-ktx')
