+++ samples/browser/src/main/java/org/mozilla/samples/browser/BrowserActivity.kt
@@ -11,7 +11,7 @@ import android.util.AttributeSet
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import mozilla.components.browser.session.intent.getSessionId
+import mozilla.components.feature.intent.ext.getSessionId
 import mozilla.components.browser.tabstray.BrowserTabsTray
 import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.tabstray.TabsTray
