+++ samples/browser/src/main/java/org/mozilla/samples/browser/DefaultComponents.kt
@@ -30,7 +30,7 @@ import mozilla.components.concept.engine.DefaultSettings
 import mozilla.components.concept.engine.Engine
 import mozilla.components.feature.customtabs.CustomTabIntentProcessor
 import mozilla.components.feature.customtabs.store.CustomTabsServiceStore
-import mozilla.components.feature.intent.TabIntentProcessor
+import mozilla.components.feature.intent.processing.TabIntentProcessor
 import mozilla.components.feature.media.MediaFeature
 import mozilla.components.feature.media.RecordingDevicesNotificationFeature
 import mozilla.components.feature.media.state.MediaStateMachine
